Major enhancements to openils.widget.PCrudFilterDialog
openils.widget.PCrudFilterDialog is (now) a dijit that allows users to define
inclusionary filters based on any number of rows that specify a field
from a given fieldmapper class, an operator, and a value (or range of
values, for the [not] between operator).  The resulting filter is
suitable for use as the where clause of a pcrud search call.

AutoGrid has an option, showLoadFilter, that allows you to easily
provide a filter dialog in any AutoGrid-based interface, with some minor
caveats (see the comments at the top of the source for
openils.widget.PCrudFilterDialog).  AutoGrid will also pass along the
suppressFilterFields option, which works analagously to the
suppressEditFields option.  You can also build your own interfaces to
use this dijit without AutoGrid, if desired.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>