LP#2049934: Remove trailing TCN spaces from incoming bib records.
For libraries using the 001 as the Evergreen database ID and the 035
to store the original title control number, this fix ensures that
any leading or trailing spaces in the incoming record's 001 are removed.
This ensures matching on existing records with the same TCN in the 035.
Authored by Galen Charlton for PINES and applied locally for many years.

Release-note: Trim spaces from TCN when importing Z39.50

Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: blake <blake@mobiusconsortium.org>
Signed-off-by: Mary Llewellyn <mllewell@biblio.org>