Default Vandelay merge profiles
Adds 2 default vandelay merge profiles, in part so users will have some
useful sample data, but also so the community can have some
basic/consistent profiles for documentation purposes.

1. Match-Only Merge

This is a no-op merge.  It's only purpose is to allow Vandelay, which
requires a profile of some sort to do anything, to find matching
records.  This is useful for importing holdings when the MARC from the
source record is not desired.  It will also be useful in the
(forthecoming) Vandelay-driven Acquisitions lineitem linking (LP
pending).

The target/incumbent record is not affected.

2. Full Overlay

Used for overlaying an entire inbound MARC record over its target
record.  The target/incumbent record is completely replaced by the
inbound record.

For both, 901c is used because it's automatically updated on the target
record to match the internal ID, regardless of any field
preservation/stripping.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>