TPAC: Cleaner title display in simple and detailed views
The TPAC uses a simple 245a display for the "simple view" in search
results, so records like the concerto.sql "Violin concerto no. 3 in G,
K. 216 ;" display with a trailing semicolon that is rather
disconcerting.

Similarly, the subfields in the extended title as displayed in the
record detail view was simply concatenating subfields together. This
resulted in titles like:

Violin concerto no. 3 in G, K. 216 ;Sinfonia ...

rather than:

Violin concerto no. 3 in G, K. 216 ; Sinfonia ...

This commit strips trailing punctuation that suggests a continuation
(:;/) from the brief title and joins the subfields with a space for the
extended title.

Also, enable the detailed view in search results to display the complete
title.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>