TPac: avoid SMS org setting check for non-SMS pages
Avoid an unnecessary org unit setting lookup on a large swath of
(mostly myopac) pages.  Instead, check to see if the current page is the
SMS page in question and the lookup is necessary before making the
lookup.

This change also inserts the physical location as the first org unit to
use in the org setting test.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>