Use getItemType() instead of getForm() in searchBarInit().
This seems to actually fix the problem with the item type limiter
being lost when going from advanced search to the results page.

Looks like the wrong method was being used to set the form_selector
value in search_bar.js. It was being set to the web form itself and
not to the value of the it (item type) parameter.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>