Vandelay record match sets: fix bug loading some trees
Reported to me by Bill Erickson.  The Vandelay Record Match Sets
interface was failing to load some simple trees of match set points.

Example: http://pastesite.com/31643

The proximate symptom was this error message:

    Error: dojo.data.ItemFileReadStore: Invalid item argument.
    Source File: http://dev198.esilibrary.com/js/dojo/dojo/dojo.js
    Line: 119

but the root cause was a problem in dojoize_match_set_tree(), which
was failing to generate unique identifiers for each item in the data
store it's responsible for generating.

Since that function is only ever used to deal with trees where all the
nodes come from the database and have actual IDs, that function now uses
the database IDs as is rather than attempting to generate new ones.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>