Acq: General Search refactor
This is a combination of the following three commits from master:

3149433d354889fb
56121886c72b2d41
35fe5b9cbe10b132

The first two were meant for backport to rel_2_0 and rel_2_1 at the
time, but the backport wasn't done.  They were buggy anyway, and now it
makes the most sense (I hope) to backport them together with the
corrective commit all at once.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>