Tpac: record detail print and email
Implements Print and Email actions (as links below add-to-list) in the
tpac record detail page.

The existing bib record print/email A/T seed templates were also updated
like so:

* Instead of using the reporter tables, bib data is fetched and
extracted via unapi, which allows access to SVF data (etc) and is more
consistent with the tpac.

* A number of unused A/T evnironment entries for the seed templates were
removed.

* The delay for the email template was explicitly set to 00:00:00
instead of the default 5 minutes to get the emails out quicker.

Signed-off-by: Bill Erickson <berick@esilibrary.com>