TPAC: serials display paging bugfix
In the previous commit, paging at the deepest level of an expanded tree
was broken.  This fixes that, and improves pager labeling a little bit.

This also makes the default page size for holdings 12 instead of 10,
since monthlies are a pretty common case, and why not show them all on
one page.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Mike Rylander <mrylander@gmail.com>