updating item out count in patron display
For checkouts (and checkouts converted into renewals), this updates the item out
count in the patron summary sidebar, and under the Items Out navigation button.

It also updates the Items Out sub-interface if that interface has been opened.

However, we are adding an extra network call to update the summary sidebar.  We
may want to have the checkout method return that data (the checkout count
summary) instead.

Signed-off-by: Jason Etheridge <jason@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>