Alt Serial Control interfaces and Batch Receive reload OPAC when possible
Really, since the Alt Serial Control interfaces are built on AutoGrid,
all we have to do there is teach AutoGrid (and its constituent EditPane)
to try for reload_opac on write operations.

This may cause more reloads of the OPAC than are truly necessary, but
mainly in cases when nobody's looking because the OPAC is hidden away
in another tab.  That's preferable to the alternative of littering the
codebase with dozens of instances of
    try { xulG.reload_opac(); } catch (E) { }
which would just be ugly, bad and unmaintainable.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>