Propagate search location on "Another Search" link
In commit 27b2037d72b I had intentionally prevented "Another Search" on
the results page from propagating the current search location, in the
belief that sites would want to use that as a way of starting a brand
new search session; after all, both the "Advanced search" and "Search
toolbar" offer propagation of the current parameters.

However, a number of people including Kathy Lussier and Ben Shum
indicated that they expected "Another Search" to maintain scope as well,
so we will propagate state here.

Sites may want to customize this link or the top-left logo link if they
want to provide a "Start a new search session" option.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>